requirements.txt
alembic==1.11.0
aniso8601==9.0.1
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Session==0.5.0
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Khayyam==3.0.17
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
smsir-python==1.0.7
SQLAlchemy==2.0.13
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
